pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opcua-testing"
version = "0.1.0"
description = "Test harness for asyncua OPC UA client/server scenarios"
readme = "README.txt"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Victor Cornejo"}]
keywords = ["opcua", "asyncua", "testing", "industrial", "iot"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Testing",
]
dependencies = [
  "asyncua>=1.0.0",
  "python-dotenv>=1.0.0",
  "aiomqtt>=2.0.0",
  "influxdb-client>=1.40.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "types-python-dateutil",
]

[tool.ruff]
line-length = 100
target-version = "py311"
fix = false
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E203"]
exclude = ["venv", ".venv", "build", "dist", "__pycache__"]

[tool.ruff.isort]
combine-as-imports = true
known-first-party = ["opcua_testing"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "asyncua.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
